Fix prune so that it doesn't cause a toolchain build failure.
TBR= robertm
Review URL: http://codereview.chromium.org/8591015

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7204 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: e89cd9f9382dec8874d47eaebf061bea862ea553

+++ pnacl/build.sh
@@ -2664,10 +2664,12 @@ prune-translator-install() {
   StepBanner "PRUNE" "Pruning translator installs [${srpc_kind}]"
   for arch in ${SBTC_BUILD_WITH_PNACL} ; do
     StepBanner "PRUNE" "llvm cruft [${arch}]"
-    spushd "${INSTALL_SB_TOOLS}/${arch}/${srpc_kind}"
-    rm -rf include lib nacl share
-    rm -rf bin/llvm-config bin/tblgen
-    spopd
+    if [ -d "${INSTALL_SB_TOOLS}/${arch}/${srpc_kind}" ]; then
+      spushd "${INSTALL_SB_TOOLS}/${arch}/${srpc_kind}"
+      rm -rf include lib nacl share
+      rm -rf bin/llvm-config bin/tblgen
+      spopd
+    fi
   done
 
   if ! ${SBTC_PRODUCTION}; then
@@ -2676,7 +2678,9 @@ prune-translator-install() {
 
   StepBanner "PRUNE" "Stripping tools-sb nexes"
   for arch in ${SBTC_BUILD_WITH_PNACL} ; do
-    ${PNACL_STRIP} "${INSTALL_SB_TOOLS}/${arch}/${srpc_kind}"/bin/*
+    if [ -d "${INSTALL_SB_TOOLS}/${arch}/${srpc_kind}" ]; then
+      ${PNACL_STRIP} "${INSTALL_SB_TOOLS}/${arch}/${srpc_kind}"/bin/*
+    fi
   done
 
   StepBanner "PRUNE" "remove driver log"
